Minimongo now uses a queue to manage observes
What this means for the user is that now, both on the server and the client, if
you look at a minimongo-backed collection from an observe/observeChanges
callback, you'll see a consistent view of it -- you'll never see it in the
middle of an update, for instance.

Furthermore, server!minimongo no longer has stubs, just future waits or
callbacks, your choice.